Fix pip package tests. (#14685)
* Disable contrib/summary tests from pip runs.

They depend on a testonly package, so they break pip test runs for
nightly and releases.

* Disable reader_ops_test from pip tests.

It depends on a testonly module, which breaks nightly and release tests.

* Instead of disabling tests, merge test_internal package into test_util.

* Add import for sqlite3 to summary_test_util.

* Add functools import to summary_test_util.
