+++ tensorflow/contrib/summary/summary_ops_graph_test.py
@@ -19,17 +19,17 @@ from __future__ import print_function
 import six
 
 from tensorflow.contrib.summary import summary_ops
-from tensorflow.contrib.summary import summary_test_internal
+from tensorflow.contrib.summary import summary_test_util
 from tensorflow.core.framework import graph_pb2
 from tensorflow.core.framework import node_def_pb2
 from tensorflow.python.framework import ops
 from tensorflow.python.platform import test
 from tensorflow.python.training import training_util
 
-get_all = summary_test_internal.get_all
+get_all = summary_test_util.get_all
 
 
-class DbTest(summary_test_internal.SummaryDbTest):
+class DbTest(summary_test_util.SummaryDbTest):
 
   def testGraphPassedToGraph_isForbiddenForThineOwnSafety(self):
     with self.assertRaises(TypeError):
